requirements.txt
Flask==2.1.0
pandas>=1.3.5
sqlalchemy>=1.4.31
datetime>=4.4
mysqlclient>=2.1.0
pymysql>=1.0.2
dash>=2.4.1
dash-table>=5.0.0
plotly>=5.8.0

